+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Added `table/options.ReadFromSnapshot(bool)` option for `session.StreamReadTable()`
+
 ## v3.51.1
 * Added checking condition for `tx.Rollback()` in `retry.DoTx`